Add files via upload
##Week 2: Diving deeper into the BBC News archive
1. Defining useful Global variables
NUM_WORDS: The maximum number of words to keep, based on word frequency. Defaults to 1000.
EMBEDDING_DIM: Dimension of the dense embedding, will be used in the embedding layer of the model. Defaults to 16.
MAXLEN: Maximum length of all sequences. Defaults to 120.
PADDING: Padding strategy (pad either before or after each sequence.). Defaults to 'post'.
OOV_TOKEN: Token to replace out-of-vocabulary words during text_to_sequence calls. Defaults to "<OOV>".
2. Loading and Pre-processing the data
3.Training- Validation split:Given the training split size, this function will split the full lists of sentences and labels into training and validation sentences and labels.
4. Tokenization- Sequences and Padding
A couple of things to note:

You should fit the tokenizer to all the labels to avoid the case of a particular label not being present in the validation set. Since you are dealing with labels there should never be an OOV label.
In the previous function you used the pad_sequences function which returns numpy arrays. Here you will not be using it since you don't need to pad the labels so you need to make the conversion to numpy arrays yourself.
The argument split_labels refers to the labels of a particular split (train or validation). This is because the function should work independently of the split being used.
Using Keras' Tokenizer yields values that start at 1 rather than at 0. This will present a problem when training since Keras usually expects the labels to start at 0. To work around this issue you could use an extra neuron in the last layer of your model. However this approach is rather hacky and not very clear. Instead you will substract 1 from every value of the labels that the function returns. Remember that when using numpy arrays you can simply do something like np.array - 1 to accomplish this since numpy allows for vectorized operations.
5. Selecting the model for text classification
For this complete the create_model below.

A couple of things to keep in mind:

Notice that this function has three parameters, all of which are meant to be passed to an Embedding layer, which is what you will probably use as a first layer for your model.
The last layer should be a Dense layer with 5 units (since there are 5 categories) with a softmax activation.
You should also compile your model using an appropiate loss function and optimizer.
You can use any architecture you want but keep in mind that this problem doesn't need many layers to be solved successfully. You don't need any layers beside Embedding, GlobalAveragePooling1D and Dense layers but feel free to try out different architectures.
To pass this graded function your model should reach at least a 95% training accuracy and a 90% validation accuracy in under 30 epochs.